code samples/max_opened_projects: refactor, cleanup
