+++ code_samples/max_opened_projects/README.md
@@ -3,16 +3,23 @@
 
 ## Quickstart
 
-Maximum Open Projects Sample implements a `ProjectManagerListener` with two methods applied to check if the current projects have been opened or closed.
-Each method refers to the `ProjectCountingService` [light service][docs:plugin_services:light_services].
-It provides methods to increase and decrease the global counter of the currently opened projects in the IDE.
-After opening each one, a message dialog is presented to the user with the current number.
+Maximum Open Projects Sample implements a `StartupActivity` extension point to run on project open as well as a
+`ProjectManagerListener` for tracking projects being closed.
+Both use `ProjectCountingService` application-level [light service][docs:plugin_services:light_services].
+It provides methods to increase and decrease the counter of currently opened projects in the IDE.
+When opening more projects than the maximum allowed (3), a message dialog is shown.
+
+### Extension Points
+
+| Name                               | Implementation                                                | Extension Point Class |
+|------------------------------------|---------------------------------------------------------------|-----------------------|
+| `com.intellij.postStartupActivity` | [ProjectOpenStartupActivity][file:ProjectOpenStartupActivity] | `StartupActivity`     |
 
 ### Application Listeners
 
-| Name     | Implementation                                            | Listener Class           |
-|----------|-----------------------------------------------------------|--------------------------|
-| listener | [ProjectOpenCloseListener][file:ProjectOpenCloseListener] | `ProjectManagerListener` |
+| Name     | Implementation                                        | Listener Class           |
+|----------|-------------------------------------------------------|--------------------------|
+| listener | [ProjectOpenCloseListener][file:ProjectCloseListener] | `ProjectManagerListener` |
 
 *Reference: [Plugin Listeners in IntelliJ SDK Docs][docs:listeners]*
 
@@ -21,5 +28,6 @@ After opening each one, a message dialog is presented to the user with the curre
 [docs:plugin_services:light_services]: https://plugins.jetbrains.com/docs/intellij/plugin-services.html#light-services
 [docs:listeners]: https://plugins.jetbrains.com/docs/intellij/plugin-listeners.html
 
+[file:ProjectOpenStartupActivity]: ./src/main/java/org/intellij/sdk/maxOpenProjects/ProjectOpenStartupActivity.java
 [file:ProjectCountingService]: ./src/main/java/org/intellij/sdk/maxOpenProjects/ProjectCountingService.java
-[file:ProjectOpenCloseListener]: ./src/main/java/org/intellij/sdk/maxOpenProjects/ProjectOpenCloseListener.java
+[file:ProjectCloseListener]: ./src/main/java/org/intellij/sdk/maxOpenProjects/ProjectCloseListener.java
