+++ topics/basics/plugin_structure/plugin_services.md
@@ -306,6 +306,6 @@ val projectService = project.service<MyProjectService>()
 To clarify how to use services, consider the **maxOpenProjects** sample plugin available in the [code samples](%gh-sdk-samples%/max_opened_projects).
 
 This plugin has an application service counting the number of currently opened projects in the IDE.
-If this number exceeds the maximum number of simultaneously opened projects allowed by the plugin (3), it displays a warning message.
+If this number exceeds the maximum number of simultaneously opened projects allowed by the plugin (3), it displays an information message.
 
 See [Code Samples](code_samples.md) on how to set up and run the plugin.
